Refactoring - fixed warnings - make some codes readable

+++ Project Ant Colony/GameSetting.cpp
@@ -6,7 +6,7 @@ namespace GameSetting
 	extern const int windowWidth{ 1500 };
 
 	extern const float GAMESPEED{ 1.0 }; //normal speed = 1.0
-	extern const float FRAMERATE{ 60 }; //default = 60fps
+	extern const int FRAMERATE{ 60 }; //default = 60fps
 	
 	extern const sf::Vector2f nullVector{ 0.0f, 0.0f };
 	extern const sf::Vector2f oneVector{ 1.0f, 1.0f };
